# contracts/m15.openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "ATLAS M15 local developer API", "version": "1.0.0", "description": "REQ-DEV-001..005. Loopback/SSH only. Stateless text/function subset of official protocols; unsupported options rejected. Cache/metadata hints confer no identity or caching guarantee. Effort hints without enabled thinking do not enable reasoning. See contracts/m15.md."},
  "servers": [{"url": "http://127.0.0.1:8030"}],
  "security": [{"bearer": []}, {"apiKey": []}],
  "paths": {
    "/v1/chat/completions": {"post": {"operationId": "create_chat_completions", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatRequest"}}}}, "responses": {"200": {"description": "Official chat.completion envelope or incremental SSE; see contracts/m15.md.", "content": {"application/json": {"schema": {"type": "object"}}, "text/event-stream": {"schema": {"type": "string"}}}}, "400": {"description": "Unsupported/invalid input, context or budget", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "401": {"description": "Missing/invalid/revoked API key", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "429": {"description": "Developer quota exhausted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "502": {"description": "Provider failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}}}},
    "/v1/responses": {"post": {"operationId": "create_responses", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResponsesRequest"}}}}, "responses": {"200": {"description": "Official response envelope or incremental SSE; see contracts/m15.md.", "content": {"application/json": {"schema": {"type": "object"}}, "text/event-stream": {"schema": {"type": "string"}}}}, "400": {"description": "Unsupported/invalid input, context or budget", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "401": {"description": "Missing/invalid/revoked API key", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "429": {"description": "Developer quota exhausted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "502": {"description": "Provider failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}}}},
    "/v1/messages": {"post": {"operationId": "create_messages", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessagesRequest"}}}}, "responses": {"200": {"description": "Official message envelope or incremental SSE; see contracts/m15.md.", "content": {"application/json": {"schema": {"type": "object"}}, "text/event-stream": {"schema": {"type": "string"}}}}, "400": {"description": "Unsupported/invalid input, context or budget", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "401": {"description": "Missing/invalid/revoked API key", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "429": {"description": "Developer quota exhausted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "502": {"description": "Provider failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}}}},
    "/v1/models": {"get": {"operationId": "list_models", "responses": {"200": {"description": "OpenAI list containing atlas-code only"}, "401": {"description": "Unauthorized"}}}},
    "/v1/usage": {"get": {"operationId": "get_usage", "responses": {"200": {"description": "Authenticated developer UTC daily counters", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Usage"}}}}, "401": {"description": "Unauthorized"}}}}
  },
  "components": {"securitySchemes": {"bearer": {"type": "http", "scheme": "bearer"}, "apiKey": {"type": "apiKey", "in": "header", "name": "x-api-key"}}, "schemas": {
    "Part": {"type": "object", "properties": {"type": {"enum": ["text", "input_text", "output_text", "tool_use", "tool_result"]}, "text": {"type": "string", "maxLength": 1048576}, "id": {"type": "string", "maxLength": 1048576}, "name": {"type": "string", "maxLength": 1048576}, "input": {"type": "object"}, "tool_use_id": {"type": "string", "maxLength": 1048576}, "content": {"type": ["string", "array"]}, "is_error": {"type": "boolean"}, "cache_control": {"type": "object"}}, "required": ["type"], "additionalProperties": false},
    "Function": {"type": "object", "properties": {"name": {"type": "string", "pattern": "^[a-zA-Z0-9_-]{1,64}$"}, "description": {"type": "string", "maxLength": 1048576}, "parameters": {"type": "object"}, "strict": {"type": "boolean"}}, "required": ["name", "parameters"], "additionalProperties": false},
    "Call": {"type": "object", "properties": {"id": {"type": "string", "maxLength": 1048576}, "type": {"const": "function"}, "function": {"type": "object", "properties": {"name": {"type": "string", "maxLength": 1048576}, "arguments": {"type": "string", "maxLength": 1048576}}, "required": ["name", "arguments"], "additionalProperties": false}}, "required": ["id", "type", "function"], "additionalProperties": false},
    "Message": {"type": "object", "properties": {"role": {"enum": ["system", "developer", "user", "assistant", "tool"]}, "content": {"oneOf": [{"type": "string", "maxLength": 1048576}, {"type": "array", "items": {"$ref": "#/components/schemas/Part"}, "maxItems": 1024}, {"type": "null"}]}, "tool_calls": {"type": "array", "items": {"$ref": "#/components/schemas/Call"}, "maxItems": 1024}, "tool_call_id": {"type": "string", "maxLength": 1048576}, "name": {"type": "string", "maxLength": 1048576}, "type": {"const": "message"}, "id": {"type": "string", "maxLength": 1048576}, "status": {"enum": ["completed", "in_progress", "incomplete"]}}, "required": ["role"], "additionalProperties": false},
    "ResponseCall": {"type": "object", "properties": {"type": {"const": "function_call"}, "call_id": {"type": "string", "maxLength": 1048576}, "name": {"type": "string", "maxLength": 1048576}, "arguments": {"type": "string", "maxLength": 1048576}, "id": {"type": "string", "maxLength": 1048576}, "status": {"enum": ["completed", "incomplete"]}}, "required": ["type", "call_id", "name", "arguments"], "additionalProperties": false},
    "ToolOutput": {"type": "object", "properties": {"type": {"const": "function_call_output"}, "call_id": {"type": "string", "maxLength": 1048576}, "output": {"type": "string", "maxLength": 1048576}, "id": {"type": "string", "maxLength": 1048576}}, "required": ["type", "call_id", "output"], "additionalProperties": false},
    "ChatRequest": {"type": "object", "properties": {"model": {"const": "atlas-code"}, "stream": {"type": "boolean", "default": false}, "tools": {"type": "array", "items": {"type": "object"}, "maxItems": 1024}, "tool_choice": {"oneOf": [{"enum": ["auto", "none", "required"]}, {"type": "object"}]}, "temperature": {"type": "number", "minimum": 0, "maximum": 2}, "messages": {"type": "array", "items": {"$ref": "#/components/schemas/Message"}, "maxItems": 1024}, "max_tokens": {"type": "integer", "minimum": 1, "maximum": 8192}, "max_completion_tokens": {"type": "integer", "minimum": 1, "maximum": 8192}, "parallel_tool_calls": {"type": "boolean"}, "stream_options": {"type": "object", "properties": {"include_usage": {"type": "boolean"}}, "required": [], "additionalProperties": false}}, "required": ["model", "messages"], "additionalProperties": false},
    "ResponsesRequest": {"type": "object", "properties": {"model": {"const": "atlas-code"}, "stream": {"type": "boolean", "default": false}, "tools": {"type": "array", "items": {"type": "object"}, "maxItems": 1024}, "tool_choice": {"oneOf": [{"enum": ["auto", "none", "required"]}, {"type": "object"}]}, "temperature": {"type": "number", "minimum": 0, "maximum": 2}, "input": {"oneOf": [{"type": "string", "maxLength": 1048576}, {"type": "array", "items": {"oneOf": [{"$ref": "#/components/schemas/Message"}, {"$ref": "#/components/schemas/ResponseCall"}, {"$ref": "#/components/schemas/ToolOutput"}]}, "maxItems": 1024}]}, "instructions": {"type": "string", "maxLength": 1048576}, "max_output_tokens": {"type": "integer", "minimum": 1, "maximum": 8192}, "parallel_tool_calls": {"type": "boolean"}, "store": {"const": false}, "reasoning": {"type": "object", "properties": {"effort": {"enum": ["none", "low"]}, "summary": {"const": "none"}}, "required": [], "additionalProperties": false}, "include": {"type": "array", "items": {"const": "reasoning.encrypted_content"}, "maxItems": 1024}, "prompt_cache_key": {"type": "string", "maxLength": 1048576}, "client_metadata": {"type": "object", "maxProperties": 32, "additionalProperties": {"type": "string", "maxLength": 1024}}, "metadata": {"type": "object", "maxProperties": 32, "additionalProperties": {"type": "string", "maxLength": 1024}}}, "required": ["model", "input"], "additionalProperties": false},
    "MessagesRequest": {"type": "object", "properties": {"model": {"const": "atlas-code"}, "stream": {"type": "boolean", "default": false}, "tools": {"type": "array", "items": {"type": "object"}, "maxItems": 1024}, "tool_choice": {"oneOf": [{"enum": ["auto", "none", "required"]}, {"type": "object"}]}, "temperature": {"type": "number", "minimum": 0, "maximum": 2}, "messages": {"type": "array", "items": {"$ref": "#/components/schemas/Message"}, "maxItems": 1024}, "system": {"oneOf": [{"type": "string", "maxLength": 1048576}, {"type": "array", "items": {"$ref": "#/components/schemas/Part"}, "maxItems": 1024}, {"type": "null"}]}, "max_tokens": {"type": "integer", "minimum": 1, "maximum": 8192}, "metadata": {"type": "object", "maxProperties": 32, "additionalProperties": {"type": "string", "maxLength": 1024}}, "thinking": {"type": "object", "properties": {"type": {"const": "disabled"}}, "required": [], "additionalProperties": false}, "output_config": {"type": "object", "properties": {"effort": {"enum": ["low", "medium", "high"]}}, "required": [], "additionalProperties": false}}, "required": ["model", "messages", "max_tokens"], "additionalProperties": false},
    "Error": {"type": "object", "required": ["error"], "properties": {"error": {"type": "object", "required": ["type", "message"], "properties": {"type": {"type": "string", "maxLength": 1048576}, "message": {"type": "string", "maxLength": 1048576}}}}},
    "Usage": {"type": "object", "properties": {"developer": {"type": "string", "maxLength": 1048576}, "day": {"type": "string", "maxLength": 1048576}, "requests": {"type": "integer"}, "charged_micro_eur": {"type": "integer"}, "unknown_micro_eur": {"type": "integer"}}, "required": ["developer", "day", "requests", "charged_micro_eur", "unknown_micro_eur"], "additionalProperties": false}
  }}
}